+++ components/rendering/src/omeis/providers/re/RenderHSBRegionTask.java
@@ -15,7 +15,7 @@ import org.slf4j.LoggerFactory;
 import ome.util.PixelData;
 import omeis.providers.re.codomain.CodomainChain;
 import omeis.providers.re.data.Plane2D;
-import omeis.providers.re.lut.LutReader;
+import omeis.providers.re.lut.LutReaderFactory;
 import omeis.providers.re.quantum.BinaryMaskQuantizer;
 import omeis.providers.re.quantum.QuantizationException;
 import omeis.providers.re.quantum.QuantumStrategy;
@@ -70,7 +70,7 @@ class RenderHSBRegionTask implements RenderingTask {
     private Optimizations optimizations;
 
     /** The collection of readers.*/
-    private List<LutReader> readers;
+    private List<LutReaderFactory> readers;
 
     /**
      * Creates a new instance to render a wavelength.
@@ -100,7 +100,7 @@ class RenderHSBRegionTask implements RenderingTask {
             List<QuantumStrategy> strategies, CodomainChain cc,
             List<int[]> colors, Optimizations optimizations,
             int x1Start, int x1End, int x2Start, int x2End,
-            List<LutReader> readers) {
+            List<LutReaderFactory> readers) {
         this.dataBuffer = dataBuffer;
         this.wData = wData;
         this.strategies = strategies;
@@ -149,7 +149,7 @@ class RenderHSBRegionTask implements RenderingTask {
         byte[] r = dataBuffer.getRedBand();
         byte[] g = dataBuffer.getGreenBand();
         byte[] b = dataBuffer.getBlueBand();
-        LutReader reader;
+        LutReaderFactory reader;
         for (Plane2D plane : wData) {
             int[] color = colors.get(i);
             reader = readers.get(i);
@@ -226,7 +226,7 @@ class RenderHSBRegionTask implements RenderingTask {
         int[] buf = ((RGBIntBuffer) dataBuffer).getDataBuffer();
         boolean isPrimaryColor = optimizations.isPrimaryColorEnabled();
         boolean isAlphaless = optimizations.isAlphalessRendering();
-        LutReader reader;
+        LutReaderFactory reader;
         for (Plane2D plane : wData) {
             int[] color = colors.get(i);
             reader = readers.get(i);
@@ -370,7 +370,7 @@ class RenderHSBRegionTask implements RenderingTask {
         int[] buf = ((RGBAIntBuffer) dataBuffer).getDataBuffer();
         boolean isPrimaryColor = optimizations.isPrimaryColorEnabled();
         boolean isAlphaless = optimizations.isAlphalessRendering();
-        LutReader reader;
+        LutReaderFactory reader;
         for (Plane2D plane : wData) {
             int[] color = colors.get(i);
             reader = readers.get(i);
